refactor(slider): move into Forms/ (@sungik-choi)
