+++ packages/bezier-react/src/components/Forms/Slider/Slider.stories.tsx
