+++ packages/bezier-react/src/components/Forms/Slider/Slider.styled.ts
