+++ packages/bezier-react/src/components/Forms/Slider/Slider.test.tsx
