+++ packages/bezier-react/src/index.ts
@@ -21,11 +21,11 @@ export * from 'Components/Forms/FormHelperText'
 export * from 'Components/Forms/FormLabel'
 export * from 'Components/Forms/Inputs/mixins'
 export * from 'Components/Forms/Inputs/Select'
-export * from 'Components/Forms/Inputs/Slider'
 export * from 'Components/Forms/Inputs/TextArea'
 export * from 'Components/Forms/Inputs/TextField'
 export * from 'Components/Forms/Radio'
 export * from 'Components/Forms/SegmentedControl'
+export * from 'Components/Forms/Slider'
 export * from 'Components/Forms/Switch'
 export * from 'Components/Icon'
 export * from 'Components/KeyValueListItem'
